requirements.txt
flask
groq
rank-bm25
numpy
python-dotenv
gunicorn